renamed errors.py and updated error messages
